refac: keywordUtil 클래스 분리
